requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
configobj==5.0.6
Django==3.0.4
djangorestframework==3.11.0
googletrans==2.4.0
idna==2.9
langcodes==2.0.0
marisa-trie==0.7.5
msgpack==1.0.0
Pillow==7.1.2
pytz==2020.1
redis==3.5.2
regex==2018.2.21
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
wincertstore==0.2
wordfreq==2.2.2
